pyproject.toml
[project]
name = "gerenciamento-consultas-medicas"
version = "0.1"
description = "API desenvolvida para teste técnico lacrei saúde"
authors = [
    {name = "akeenathon",email = "akeenathon@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref (==3.9.0)",
    "django (==5.2.4)",
    "django-cors-headers (==4.7.0)",
    "djangorestframework (==3.16.0)",
    "djangorestframework-simplejwt (==5.5.0)",
    "pyjwt (==2.9.0)",
    "sqlparse (==0.5.3)",
    "tzdata (==2025.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
